I liked fat models more anyways
