+++ app/models/user.rb
@@ -1,5 +1,4 @@
 class User < ActiveRecord::Base
-  include Usable
   acts_as_paranoid
 
   devise :database_authenticatable, :registerable, :confirmable,
@@ -64,6 +63,42 @@ class User < ActiveRecord::Base
     attendances_for(TeaTime.future).each do |a|
       a.flake!(email: false)
     end
+  end
+
+  # Devise Methods
+  def password_required?
+    super if confirmed?
+    # Password is required if it is being set, but not for new records
+    if !persisted?
+      false
+    else
+      !password.nil? || !password_confirmation.nil?
+    end
+  end
+
+  def password_match?
+    self.errors[:password] << "can't be blank" if password.blank?
+    self.errors[:password_confirmation] << "can't be blank" if password_confirmation.blank?
+    self.errors[:password_confirmation] << "does not match password" if password != password_confirmation
+    password == password_confirmation && !password.blank?
+  end
+
+  # new function to set the password without knowing the current password used in our confirmation controller.
+  def attempt_set_password(params)
+    p = {}
+    p[:password] = params[:password]
+    p[:password_confirmation] = params[:password_confirmation]
+    update_attributes(p)
+  end
+
+  # new function to return whether a password has been set
+  def has_no_password?
+    self.encrypted_password.blank?
+  end
+
+  # new function to provide access to protected method unless_confirmed
+  def only_if_unconfirmed
+    pending_any_confirmation { yield }
   end
 
   class << self
